backend/requirements.txt
fastapi
langchain_groq
crewaiuvicorn
pydantic
python-dotenv